+++ CHANGELOG.md
@@ -35,6 +35,7 @@ This release introduces `configVersion: 2` with clearer rules for the configurat
 * **New** `bypassModes` setting in defaults of the config, where the user can set modes where Macrowhisper should not kick in at all.
   * Useful since SuperWhisper now allows overriding the auto-paste setting at the mode level.
   * It is still suggested that users set Superwhisper's autopaste off in the advanced configuration tab. However, if they do want to use SuperWhisper for pasting with a specific mode, now they can bypass that specifically with Macrowhisper. No actions will trigger when using that mode.
+* **New** `{{appVocabulary}}` which captures names, usernames, and special terms from the active window.
 * **Improvement.** `inputCondition` has been expanded to all action types
   * You can now have any action behave differently depending on user being in an input field or not.
 * **Improvement.** Streamlined the validation and sync with Superwhisper's placement of the result on the user's clipboard before action execution. This improves responsiveness.